refactor(feedback): extract shared form logic, centralize contact info, and polish reporting flow
