+++ src/config/constants.ts
@@ -1,0 +1,2 @@
+export const JAWAFDEHI_WHATSAPP_NUMBER = "+1 (206) 530-9098";
+export const JAWAFDEHI_EMAIL = "inquiry@jawafdehi.org";
