+++ src/i18n/locales/en.json
@@ -113,8 +113,8 @@
     "version": "Version",
     "failedToLoad": "Failed to load case details",
     "notFound": "Case not found",
-    "reportInfo": "Report missing/wrong information",
-    "reportInfoDesc": "If you find any inaccurate or missing information in this case, please let us know.",
+    "reportInfo": "Submit New Info / Request Correction",
+    "reportInfoDesc": "Help us keep our information accurate. Use this form to submit new evidence or request a correction for this case.",
     "editCase": "Edit Case",
     "contact": "Contact Us",
     "contactDesc": "You can also reach us via email or WhatsApp for any inquiries or to provide additional information.",
@@ -350,7 +350,7 @@
     "whatsappPlaceholder": "+977 98XXXXXXXX",
     "instagram": "Instagram",
     "facebook": "Facebook",
-    "other": "Other",
+    "otherContactMethod": "Other",
     "otherPlaceholder": "Enter contact details",
     "importantNotice": "Important Notice",
     "importantNoticeText": "All submissions are reviewed by our moderation team for accuracy and credibility. False allegations may have legal consequences. Please ensure your report is factual and evidence-based.",
@@ -375,7 +375,7 @@
   "feedback": {
     "title": "Platform Feedback",
     "addContactMethod": "Add Contact",
-    "description": "Help us improve Jawafdehi. Your feedback is valuable to us.",
+    "titleDescription": "Help us improve Jawafdehi. Your feedback is valuable to us.",
     "feedbackType": "Feedback Type",
     "selectFeedbackType": "Select feedback type",
     "bug": "Bug Report",
@@ -385,7 +385,7 @@
     "general": "General Feedback",
     "subject": "Subject",
     "subjectPlaceholder": "Brief summary of your feedback",
-    "description": "Description",
+    "descriptionLabel": "Description",
     "descriptionPlaceholder": "Please provide detailed feedback. For bugs, include steps to reproduce the issue.",
     "relatedPage": "Related Page or Feature (optional)",
     "relatedPagePlaceholder": "e.g., Cases page, Search functionality",
@@ -401,7 +401,7 @@
     "whatsappPlaceholder": "+977 98XXXXXXXX",
     "instagram": "Instagram",
     "facebook": "Facebook",
-    "other": "Other",
+    "otherContactMethod": "Other",
     "otherPlaceholder": "Enter contact details",
     "submitFeedback": "Submit Feedback",
     "submitting": "Submitting...",
