+++ src/i18n/locales/ne.json
@@ -113,8 +113,8 @@
     "version": "संस्करण",
     "failedToLoad": "मुद्दा विवरण लोड गर्न असफल भयो",
     "notFound": "मुद्दा फेला परेन",
-    "reportInfo": "गलत/छुटेको जानकारी रिपोर्ट गर्नुहोस्",
-    "reportInfoDesc": "यदि तपाईंले यस मुद्दामा कुनै गलत वा छुटेको जानकारी फेला पार्नुभयो भने, कृपया हामीलाई जानकारी दिनुहोस्।",
+    "reportInfo": "नयाँ जानकारी पठाउनुहोस् वा सुधार अनुरोध गर्नुहोस्",
+    "reportInfoDesc": "हाम्रो जानकारी सही राख्न मद्दत गर्नुहोस्। यस मुद्दाको लागि नयाँ प्रमाण पेश गर्न वा सुधार अनुरोध गर्न यो फारम प्रयोग गर्नुहोस्।",
     "editCase": "मुद्दा सम्पादन गर्नुहोस्",
     "contact": "हामीलाई सम्पर्क गर्नुहोस्",
     "contactDesc": "तपाईं हामीलाई कुनै पनि सोधपुछ वा थप जानकारी प्रदान गर्न इमेल वा व्हाट्सएप मार्फत पनि सम्पर्क गर्न सक्नुहुन्छ।",
@@ -350,7 +350,7 @@
     "whatsappPlaceholder": "+977 98XXXXXXXX",
     "instagram": "इन्स्टाग्राम",
     "facebook": "फेसबुक",
-    "other": "अन्य",
+    "otherContactMethod": "अन्य",
     "otherPlaceholder": "सम्पर्क विवरण प्रविष्ट गर्नुहोस्",
     "importantNotice": "महत्वपूर्ण सूचना",
     "importantNoticeText": "सबै पेशहरू सटीकता र विश्वसनीयताको लागि हाम्रो मध्यस्थता टोलीद्वारा समीक्षा गरिन्छ। झूटा आरोपहरूका कानूनी परिणामहरू हुन सक्छन्। कृपया तपाईंको रिपोर्ट तथ्यात्मक र प्रमाण-आधारित छ भनेर सुनिश्चित गर्नुहोस्।",
@@ -374,8 +374,7 @@
   },
   "feedback": {
     "title": "प्लेटफर्म प्रतिक्रिया",
-    "description": "सार्वजनिक जवाफदेहिता प्लेटफर्मलाई सुधार गर्न हामीलाई मद्दत गर्नुहोस्। तपाईंको प्रतिक्रिया हाम्रो लागि मूल्यवान छ।",
-    "addContactMethod": "सम्पर्क माध्यम थप्नुहोस्",
+    "titleDescription": "सार्वजनिक जवाफदेहिता प्लेटफर्मलाई सुधार गर्न हामीलाई मद्दत गर्नुहोस्। तपाईंको प्रतिक्रिया हाम्रो लागि मूल्यवान छ।",
     "feedbackType": "प्रतिक्रिया प्रकार",
     "selectFeedbackType": "प्रतिक्रिया प्रकार छान्नुहोस्",
     "bug": "बग रिपोर्ट",
@@ -385,7 +384,7 @@
     "general": "सामान्य प्रतिक्रिया",
     "subject": "विषय",
     "subjectPlaceholder": "तपाईंको प्रतिक्रियाको संक्षिप्त सारांश",
-    "description": "विवरण",
+    "descriptionLabel": "विवरण",
     "descriptionPlaceholder": "कृपया विस्तृत प्रतिक्रिया प्रदान गर्नुहोस्। बगहरूको लागि, समस्यालाई पुनरुत्पादन गर्नका लागि चरणहरू समावेश गर्नुहोस्।",
     "relatedPage": "सम्बन्धित पृष्ठ वा फिचर (वैकल्पिक)",
     "relatedPagePlaceholder": "जस्तै, मुद्दाहरू पृष्ठ, खोज कार्यक्षमता",
@@ -401,7 +400,7 @@
     "whatsappPlaceholder": "+977 98XXXXXXXX",
     "instagram": "इन्स्टाग्राम",
     "facebook": "फेसबुक",
-    "other": "अन्य",
+    "otherContactMethod": "अन्य",
     "otherPlaceholder": "सम्पर्क विवरण प्रविष्ट गर्नुहोस्",
     "submitFeedback": "प्रतिक्रिया पेश गर्नुहोस्",
     "submitting": "पेश गर्दै...",
